move icons to one file
