+++ src/Components/Confirmation/Confirmation.js
@@ -19,15 +19,15 @@ import { getStepDirectory, getStepNumber, STARTING_QUESTION_NUMBER } from '../..
 import { useContext, useEffect } from 'react';
 import { Context } from '../Wrapper/Wrapper.tsx';
 import Grid from '@mui/material/Grid';
-import { ReactComponent as Edit } from '../../Assets/edit.svg';
-import { ReactComponent as Residence } from '../../Assets/residence.svg';
-import { ReactComponent as Household } from '../../Assets/household.svg';
-import { ReactComponent as Head } from '../../Assets/head.svg';
-import { ReactComponent as Expenses } from '../../Assets/expenses.svg';
-import { ReactComponent as Resources } from '../../Assets/resources.svg';
-import { ReactComponent as Benefits } from '../../Assets/benefits.svg';
-import { ReactComponent as Immediate } from '../../Assets/immediate.svg';
-import { ReactComponent as Referral } from '../../Assets/referral.svg';
+import { ReactComponent as Edit } from '../../Assets/icons/edit.svg';
+import { ReactComponent as Residence } from '../../Assets/icons/residence.svg';
+import { ReactComponent as Household } from '../../Assets/icons/household.svg';
+import { ReactComponent as Head } from '../../Assets/icons/head.svg';
+import { ReactComponent as Expenses } from '../../Assets/icons/expenses.svg';
+import { ReactComponent as Resources } from '../../Assets/icons/resources.svg';
+import { ReactComponent as Benefits } from '../../Assets/icons/benefits.svg';
+import { ReactComponent as Immediate } from '../../Assets/icons/immediate.svg';
+import { ReactComponent as Referral } from '../../Assets/icons/referral.svg';
 import './Confirmation.css';
 import PreviousButton from '../PreviousButton/PreviousButton';
 import { House } from '@mui/icons-material';
